Create print elements of linked list
This challenge is part of a MyCodeSchool tutorial track and is accompanied by a video lesson.

This is an to practice traversing a linked list. Given a pointer to the head node of a linked list, print each node's  element, one per line. If the head pointer is null (indicating the list is empty), there is nothing to print